+++ packages/ice/src/constant.ts
@@ -52,6 +52,7 @@ export const RUNTIME_EXPORTS = [
       'useNavigate',
       'useNavigation',
       'useRevalidator',
+      'useAsyncValue',
     ],
     source: '@ice/runtime/router',
   },